+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Unreleased
 
+* Update StoppedMockClock to support newer phpunit (use namespaced assert class) and add
+  unit tests.
+
 ### v0.1.6 (2018-09-06)
 
 * Add AbstractArrayRepository